add wallpapers

+++ update.sh
@@ -1,6 +1,11 @@
 set -e
 
 git fetch --depth 1
-git reset --hard origin/main
-git clean -dfx
-git gc --prune=all
+if [[ $(git rev-parse HEAD) == $(git rev-parse @{u}) ]]
+then
+	echo "already up to date"
+else
+	git reset --hard origin/main
+	git clean -dfx
+	git gc --prune=all
+fi
